Updated "recent updates"  (#803)
* Change in head and subtitle

Moved the header to the top and changed subtitles.

* Updated project_activations.html

Updated to the Disaster Services page to connect to the Event timeline archive and feed from the DST markdown posts.

* Added Activation Images

Added the images for the new Disaster Services layout

* Updated Project Activations

* deleted unnecessary images

deleted images wont be used anymore. They are now available through CDN

* Update project_activations.html

* Update project_activations.html

* Update to Disaster Feed

The feed now includes icons to make it more digestible. Each Icon represents the type of event.

Co-authored-by: Siteleaf <bot@siteleaf.com>